Extract dispatcher logic from NerdTree into `AbstractDispatcher`
